requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.0.0
pymongo==4.6.0
pillow>=9.0.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0
datetime
json5==0.9.14 